Creating a hypothesis to start off
Surveying the code led me to reread the progress report, Chapter4Exercise11 as part of the SQ3R i had only one question in mind, how or if there was  if statements and for loops nested within each other that allows the one loop of a vector array for every vector element of another loop, in my previous notes it reads and i quote 
"While within the code the for-loop statement (int = 1; i< prime.size(); ++i) is wrapped around for loop that incrementally goes from 1 to 100, which I expect for each vector element of the prime.size the for loop has looped up to 100 times and nested in the  if statement “for (int i = 0; i < 100; ++i)” to with the condition that states for each value of “for (int i = 0; i < 100; ++i)” to be modulo with the current prime number that is incrementally going up the vector element of  the prime list using the for loop statement, (int = 1; i< prime.size(); ++i)." 
So i believe currently that the n for loop is the for loop suggested in my previous notes that is wrapped around the prime.size for loop, making every vector element of the n loop interact with the prime.size array, that is what I want so I can multiply current prime of each number of all_numbers but once I add the result to the chop_list after all the multiplication is done how do I remove each entry in chop_list from all_numbers?
here is my assumption or, possible axiom
There is an implication in the question when mentioned that" i put the results of the currentprime by the number provided from the n for loop, insides vector array denoted as all_numbers?" Thats what i was thinking before but now i think of it as its asking me to delete the entry (the results from the multiplication) in the choplist from all numbers, assuming that all numbers literally means all the numbers from 1 to 100 which i thought all_numbers  as my n for loop. 
Heres another thought of mine, assuming that the code is working properly when it comes to "multiply current prime of each number of all_numbers, add the result to the choplist after all the multiplication is done, 
how do i remove each entry in chop_list from all_numbers and empty choplist while leave the prime numbers in the prime list?